added graphql mock binary
