+++ backend/Cargo.toml
@@ -17,4 +17,9 @@ uuid = "1.4.0"
 axum = {version = "0.6.18", features = ["http2", "macros"]}
 async-graphql = "5.0.10"
 async-graphql-axum = "5.0.10"
-tokio = {version = "1.29.0", features = ["macros"]}
+tokio = {version = "1.29.0", features = ["full"]}
+
+
+[[bin]]
+name = "graphql_mock"
+path = "src/layer/trigger/graphql/bin.rs"
